Cambiar el tipo de dato de IdItem de Short a Int, estaba mal creado

+++ App/AutomataExistencias/AutomataExistencias.DataAccess/Aldebaran/Packaging.cs
@@ -14,7 +14,7 @@ namespace AutomataExistencias.DataAccess.Aldebaran
         public short PackagingId { get; set; }
 
         [Column("IDITEM")]
-        public short? ItemId { get; set; }
+        public int? ItemId { get; set; }
 
         [Column("PESO")]
         public double? Weight { get; set; }
